requirements.txt
Babel==1.3
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
eventlet==0.15.2
iso8601==0.1.10
kombu==3.0.23
netaddr==0.7.12
oslo.config==1.4.0
oslo.i18n==1.0.0
oslo.messaging==1.4.1
oslo.utils==1.0.0
pytz==2014.7
six==1.8.0
stevedore==1.0.0
wsgiref==0.1.2
